# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /revenue:
    get:
      operationId: listRevenue
      tags: [revenue]
      summary: List revenue entries
      parameters:
        - name: source
          in: query
          schema:
            type: string
            enum: [ads, youtube, subscriptions, apps, freelance, other]
        - name: projectId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Revenue entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RevenueEntry"
    post:
      operationId: createRevenue
      tags: [revenue]
      summary: Add a revenue entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRevenueBody"
      responses:
        "201":
          description: Created revenue entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RevenueEntry"
  /revenue/{id}:
    patch:
      operationId: updateRevenue
      tags: [revenue]
      summary: Update a revenue entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRevenueBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RevenueEntry"
    delete:
      operationId: deleteRevenue
      tags: [revenue]
      summary: Delete a revenue entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /tasks:
    get:
      operationId: listTasks
      tags: [tasks]
      summary: List all tasks
      parameters:
        - name: priority
          in: query
          schema:
            type: string
            enum: [low, medium, high, urgent]
        - name: completed
          in: query
          schema:
            type: boolean
        - name: projectId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
    post:
      operationId: createTask
      tags: [tasks]
      summary: Create a new task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTaskBody"
      responses:
        "201":
          description: Created task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /tasks/{id}:
    patch:
      operationId: updateTask
      tags: [tasks]
      summary: Update a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTaskBody"
      responses:
        "200":
          description: Updated task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
    delete:
      operationId: deleteTask
      tags: [tasks]
      summary: Delete a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /tasks/{id}/toggle:
    patch:
      operationId: toggleTask
      tags: [tasks]
      summary: Toggle task completion
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Toggled task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
...
  /wins:
    get:
      operationId: listWins
      tags: [wins]
      summary: List recent wins (newest first)
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Wins
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Win"
    post:
      operationId: createWin
      tags: [wins]
      summary: Capture a win
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWinBody"
      responses:
        "201":
          description: Created win
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Win"
  /wins/{id}:
    delete:
      operationId: deleteWin
      tags: [wins]
      summary: Delete a win
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    RevenueEntry:
      type: object
      properties:
        id:
          type: integer
        amount:
          type: number
        source:
          type: string
          enum: [ads, youtube, subscriptions, apps, freelance, other]
        description:
          type: ["string", "null"]
        projectId:
          type: ["integer", "null"]
        date:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
      required: [id, amount, source, date, createdAt]
    CreateRevenueBody:
      type: object
      properties:
        amount:
          type: number
        source:
          type: string
          enum: [ads, youtube, subscriptions, apps, freelance, other]
        description:
          type: ["string", "null"]
        projectId:
          type: ["integer", "null"]
        date:
          type: string
          format: date-time
      required: [amount, source, date]
    UpdateRevenueBody:
      type: object
      properties:
        amount:
          type: number
        source:
          type: string
          enum: [ads, youtube, subscriptions, apps, freelance, other]
        description:
          type: ["string", "null"]
        projectId:
          type: ["integer", "null"]
        date:
          type: string
          format: date-time
...
    Task:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        priority:
          type: string
          enum: [low, medium, high, urgent]
        completed:
          type: boolean
        projectId:
          type: ["integer", "null"]
        dueDate:
          type: ["string", "null"]
          format: date-time
        completedAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
      required: [id, title, priority, completed, createdAt]
    CreateTaskBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: ["string", "null"]
        priority:
          type: string
          enum: [low, medium, high, urgent]
          default: medium
        projectId:
          type: ["integer", "null"]
        dueDate:
          type: ["string", "null"]
          format: date-time
      required: [title]
    UpdateTaskBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: ["string", "null"]
        priority:
          type: string
          enum: [low, medium, high, urgent]
        completed:
          type: boolean
        projectId:
          type: ["integer", "null"]
        dueDate:
          type: ["string", "null"]
          format: date-time
...
    Win:
      type: object
      properties:
        id:
          type: integer
        text:
          type: string
        projectId:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
      required: [id, text, createdAt]
    CreateWinBody:
      type: object
      properties:
        text:
          type: string
          minLength: 1
        projectId:
          type: ["integer", "null"]
      required: [text]
...